requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
langchain-core==0.1.10

# Vector store and embeddings
chromadb==0.4.22
openai==1.10.0

# Document processing with OCR support
docling==1.0.0
pypdf==3.17.4

# OCR engines (choose one or more)
easyocr==1.7.0  # Recommended - best accuracy
# pytesseract==0.3.10  # Alternative - needs tesseract installed
# rapidocr-onnxruntime==1.3.11  # Lightweight option

# Hybrid retrieval (BM25)
rank-bm25==0.2.2

# Web interface
streamlit==1.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Graph workflow
langgraph==0.0.20

# Additional dependencies
tiktoken==0.5.2

# Image processing for OCR
pillow==10.1.0
opencv-python==4.8.1.78